[2016.11.14] - Restructuring;  - Organizing the server side code in proper directory structure;
